Create OOPs Assignment -cart
Create a class Product with two attributes- 1. productName- String and price- int.
Create an abstract class Category which has two methods- 1. getProducts and addProducts.
After that, create another class Cart which implements Category class. The cart must contain a list of Products. Initially this list is empty.
Override the methods from the Category class inside Cart. The getProducts method must return the current list of products in the cart.
Similarly, the addProduct method must add a Product to the existing list of products.
During execution, create an instance of Cart and print the list of products using the appropriate method.
After that, add two or three new products in the cart using the appropriate method and print the cart again.